maildirsizes tool to calculate sizes of maildirs without du.
Primarily because when using glusterfs as in our case, the total count
of files gets to be a problem, we need a reliable way to extract this
information on a regular basis without having to go through du and other
tools which really kills IO capacity on glsuterfs.

Having the total disk size available is useful for calculations and for
giving users reports as to their overall disk storage requirements,
again without du which can take a very, very long time when you have
users with several hundred thousand messages in a folder.